requirements.txt
ultralytics>=8.0.0
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.8.0
pandas>=2.0.0
pyyaml>=6.0
onnxruntime>=1.16.0
fastapi>=0.95.0
uvicorn>=0.22.0
requests>=2.31.0
